Finish question before 201

+++ LeetCode/165.cpp
@@ -11,7 +11,7 @@ class Solution {
 public:
     int compareVersion(string version1, string version2) {
         int i = 0, j = 0;
-        int n1 = version1.size(), n2 = version2.size();
+        int n1 = (int)version1.size(), n2 = (int)version2.size();
         int num1 = 0, num2 = 0;
         while (i < n1 || j < n2) {
             while (i < n1 && version1[i] != '.') num1 = num1 * 10 + version1[i++] - '0';
